refactor(adapters): isolate protocol state and close focus flake
