+++ server/adapters/codex-ledger.ts
@@ -1,0 +1,113 @@
+// The adapter's ledger against Codex's protocol digest
+// (codex-protocol.digest.json, from `codex app-server generate-json-schema`;
+// codex-protocol.test.ts holds them equal). Every kind the engine can send is
+// either handled by the mapper (codex-events.ts), deliberately ignored here
+// with its reason, or reported by UnknownKindReporter the first time it
+// arrives.
+export const CODEX_HANDLED_ITEMS = [
+  "agentMessage",
+  "reasoning",
+  "commandExecution",
+  "fileChange",
+  "mcpToolCall",
+  "webSearch",
+  "contextCompaction",
+  "plan",
+  "enteredReviewMode",
+  "exitedReviewMode",
+  "collabAgentToolCall",
+  "subAgentActivity",
+  "imageView",
+  "imageGeneration",
+  "dynamicToolCall",
+  "sleep",
+] as const;
+export const CODEX_IGNORED_ITEMS: Record<string, string> = {
+  userMessage: "the engine's echo of the prompt; the registry already emitted user_prompt",
+  functionCallOutput: "the raw output of a call already represented by its own item",
+  hookPrompt: "a hook's prompt fragments are the engine's input, not its output",
+};
+export const CODEX_HANDLED_METHODS = [
+  "turn/started",
+  "turn/completed",
+  "item/started",
+  "item/completed",
+  "item/agentMessage/delta",
+  "item/plan/delta",
+  "item/commandExecution/outputDelta",
+  "item/fileChange/outputDelta",
+  "item/reasoning/summaryTextDelta",
+  "item/reasoning/textDelta",
+  "turn/plan/updated",
+  "thread/tokenUsage/updated",
+  "error",
+  "warning",
+  "deprecationNotice",
+  "configWarning",
+  "guardianWarning",
+  "model/rerouted",
+] as const;
+export const CODEX_IGNORED_METHODS: Record<string, string> = {
+  "thread/started": "the session consumes it on thread/start",
+  "thread/status/changed": "the registry derives status from the stream itself",
+  "thread/name/updated": "Mirafold names sessions by folder",
+  "thread/goal/updated": "goal bookkeeping for Codex's own UI",
+  "thread/goal/cleared": "goal bookkeeping for Codex's own UI",
+  "thread/queue/changed": "Mirafold serializes prompts itself",
+  "thread/archived": "thread lifecycle for Codex's own UI",
+  "thread/unarchived": "thread lifecycle for Codex's own UI",
+  "thread/deleted": "thread lifecycle for Codex's own UI",
+  "thread/closed": "thread lifecycle for Codex's own UI",
+  "thread/reverted": "thread lifecycle for Codex's own UI",
+  "thread/project/updated": "project bookkeeping for Codex's own UI",
+  "thread/settings/updated": "settings echo; the adapter owns its /model and /effort state",
+  "thread/environment/connected": "remote-environment plumbing Mirafold does not use",
+  "thread/environment/disconnected": "remote-environment plumbing Mirafold does not use",
+  "project/changed": "project bookkeeping for Codex's own UI",
+  "skills/changed": "the prompt-option refresh re-lists skills per turn",
+  "hook/started": "Codex hooks run silently in the terminal too",
+  "hook/completed": "Codex hooks run silently in the terminal too",
+  "item/autoApprovalReview/started": "internal review of an auto-approval; the approval itself is surfaced",
+  "item/autoApprovalReview/completed": "internal review of an auto-approval; the approval itself is surfaced",
+  "autoApprovalReview/strictReviewRequired": "internal review of an auto-approval; the approval itself is surfaced",
+  "item/reasoning/summaryPartAdded": "a paragraph boundary inside reasoning already streamed as deltas",
+  "item/mcpToolCall/progress": "progress ticks of a call whose completion is shown",
+  "serverRequest/resolved": "the answer to an ask the session itself resolved",
+  "command/exec/outputDelta": "the exec runtime's own streaming; the item completion is shown",
+  "process/outputDelta": "background process plumbing not represented in the transcript",
+  "process/exited": "background process plumbing not represented in the transcript",
+  "item/commandExecution/terminalInteraction": "interactive-terminal plumbing not represented in the transcript",
+  "mcpServer/oauthLogin/completed": "MCP server administration",
+  "mcpServer/startupStatus/updated": "MCP server administration",
+  "mcpServer/event/stream/notification": "MCP server administration",
+  "account/updated": "account administration",
+  "account/login/completed": "account administration",
+  "app/list/updated": "Codex apps administration",
+  "remoteControl/status/changed": "remote-control administration",
+  "externalAgentConfig/import/progress": "config import administration",
+  "externalAgentConfig/import/completed": "config import administration",
+  "fs/changed": "the Changes panel watches the tree itself",
+  "fuzzyFileSearch/sessionUpdated": "Codex's own file picker",
+  "fuzzyFileSearch/sessionCompleted": "Codex's own file picker",
+  "thread/realtime/started": "voice/realtime mode Mirafold does not drive",
+  "thread/realtime/itemAdded": "voice/realtime mode Mirafold does not drive",
+  "thread/realtime/item/started": "voice/realtime mode Mirafold does not drive",
+  "thread/realtime/item/transcript/delta": "voice/realtime mode Mirafold does not drive",
+  "thread/realtime/item/completed": "voice/realtime mode Mirafold does not drive",
+  "thread/realtime/transcript/delta": "voice/realtime mode Mirafold does not drive",
+  "thread/realtime/transcript/done": "voice/realtime mode Mirafold does not drive",
+  "thread/realtime/outputAudio/delta": "voice/realtime mode Mirafold does not drive",
+  "thread/realtime/sdp": "voice/realtime mode Mirafold does not drive",
+  "thread/realtime/error": "voice/realtime mode Mirafold does not drive",
+  "thread/realtime/closed": "voice/realtime mode Mirafold does not drive",
+  "windows/worldWritableWarning": "Windows-only setup diagnostics",
+  "windowsSandbox/setupCompleted": "Windows-only setup diagnostics",
+  "turn/moderationMetadata": "moderation bookkeeping with no user-facing text",
+  "model/safetyBuffering/updated": "moderation bookkeeping with no user-facing text",
+  "model/verification": "model bookkeeping with no user-facing text",
+  "thread/compacted": "the contextCompaction item is the shown form of the same event",
+  "item/fileChange/patchUpdated": "the completed fileChange item carries the final changes (TS.6)",
+  "turn/diff/updated": "the Changes panel watches the working tree itself",
+  "account/rateLimits/updated":
+    "rate-limit bookkeeping on every turn; an 'approaching the limit' notice like Claude's needs the params shape recorded first",
+};
